Move KahanSum to Math/Util.h
