+++ math/mathcore/inc/Math/Math.h
@@ -139,39 +139,6 @@ inline double expm1( double x) {
 #endif
 }
 
-      template<class T>
-      class KahanSum {
-      public:
-         void Add(const T &x)
-         {
-            auto y = x - correction;
-            auto t = sum + y;
-            correction = (t - sum) - y;
-            sum = t;
-         }
-
-         void Add(const std::vector<T> &elements)
-         {
-            for (auto e : elements)
-               this->Add(e);
-         }
-
-         static double Accumulate(const std::vector<T> &elements)
-         {
-            KahanSum init;
-            init.Add(elements);
-            return init.sum;
-         }
-
-         double Result()
-         {
-            return sum;
-         }
-
-      private:
-         double sum = 0.;
-         double correction = 0.;
-      };
    } // end namespace Math
 
 } // end namespace ROOT
